+++ NuGetVersionChecker/GlobalSuppressions.cs
@@ -1,0 +1,14 @@
+﻿// This file is used by Code Analysis to maintain SuppressMessage
+// attributes that are applied to this project.
+// Project-level suppressions either have no target or are given
+// a specific target and scoped to a namespace, type, member, etc.
+
+using System.Diagnostics.CodeAnalysis;
+
+[assembly: SuppressMessage("Style", "IDE0270:Use coalesce expression", Justification = "<Pending>", Scope = "member", Target = "~M:NuGetVersionChecker.NuGetVersionChecker.GetPackageFromNuGet.GetPackageFromNugetAsync(System.String)~System.Threading.Tasks.Task{NuGetVersionChecker.NugetVersionChecker.Package}")]
+[assembly: SuppressMessage("Style", "IDE0270:Use coalesce expression", Justification = "<Pending>", Scope = "member", Target = "~M:NuGetVersionChecker.NuGetVersionChecker.GetPackageFromNuGet.GetPackagesFromNugetAsync(System.Collections.Generic.List{System.String})~System.Threading.Tasks.Task{System.Collections.Generic.List{NuGetVersionChecker.NugetVersionChecker.Package}}")]
+[assembly: SuppressMessage("Style", "IDE0090:Use 'new(...)'", Justification = "<Pending>", Scope = "member", Target = "~M:NuGetVersionChecker.NuGetVersionChecker.GetPackageFromNuGet.GetPackageFromNugetAsync(System.String)~System.Threading.Tasks.Task{NuGetVersionChecker.NugetVersionChecker.Package}")]
+[assembly: SuppressMessage("Style", "IDE0090:Use 'new(...)'", Justification = "<Pending>", Scope = "member", Target = "~M:NuGetVersionChecker.NuGetVersionChecker.GetPackageFromNuGet.GetPackagesFromNugetAsync(System.Collections.Generic.List{System.String})~System.Threading.Tasks.Task{System.Collections.Generic.List{NuGetVersionChecker.NugetVersionChecker.Package}}")]
+[assembly: SuppressMessage("Style", "IDE0090:Use 'new(...)'", Justification = "<Pending>", Scope = "member", Target = "~M:NuGetVersionChecker.NuGetVersionChecker.GetPackages(System.String)~System.Collections.Generic.List{NuGetVersionChecker.NugetVersionChecker.Package}")]
+[assembly: SuppressMessage("Usage", "CA2208:Instantiate argument exceptions correctly", Justification = "<Pending>", Scope = "member", Target = "~M:NuGetVersionChecker.NuGetVersionChecker.GetPackageFromNuGet.GetPackageFromNugetAsync(System.String)~System.Threading.Tasks.Task{NuGetVersionChecker.NugetVersionChecker.Package}")]
+[assembly: SuppressMessage("Usage", "CA2208:Instantiate argument exceptions correctly", Justification = "<Pending>", Scope = "member", Target = "~M:NuGetVersionChecker.NuGetVersionChecker.GetPackageFromNuGet.GetPackagesFromNugetAsync(System.Collections.Generic.List{System.String})~System.Threading.Tasks.Task{System.Collections.Generic.List{NuGetVersionChecker.NugetVersionChecker.Package}}")]